LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-transition-002-expected.txt


FAIL Transition with anchor names defined in different tree scopes assert_equals: expected 150 but got 188
